Completed the single_decay API test
Here the API is compared with parsing of a model specification (from a string which can come from a file as well).

What is still needed is the completion of the return methods of the Result class, as well as some diagnostic information (in the form of a report with plots and a printout of the parameters with standard errors).