Changed order of authors to signifiy Joern's contributions as the main development lead.
Updated setup.py and sphinx documentation used conf.py with changed author ordering to reflect in auto generated documentation.

As the creator of the original (Java based) glotaran software package and the first author of the first glotaran publication [1] I would like to recognize the efforts of Joern Weissenborn (@joernweissenborn) in particular in creating this Python implementation from scratch by listing him first in the author's list. This incarnation of the glotaran software, pyglotaran if you will, also includes a computational core previously only available in the form of the open source R package TIMP [2], so it is fully self contained. This incarnation of the software is the future of glotaran and perhaps even global and target analysis software in general and will eventually (soon) replace the original Glotaran software as the only supported version of glotaran going forward.

- Joris Snellenburg (@jsnel)

[1] Snellenburg, Joris, et al. "Glotaran: a Java-based graphical user interface for the R package TIMP." Journal of Statistical Software 49.3 (2012).
[2] Mullen, Katharine M., and Ivo HM Van Stokkum. "TIMP: an R package for modeling multi-way spectroscopic measurements." (2007).