Fix CODEOWNERS file (#508)
* Changed @pyglotaran_creators to @glotaran/pyglotaran_creators

* Added missing trailing 's' to file_formats path

* Exploded pattern for tooling and docs, to have the proper formatting

Before filepatterns were interpreted as owners, which lead to a failing check

* Commented out pattern for *.md files in the docs for now

Right now we don't have any *.md in the docs folder, thus this causes an error in the 'File Exist Checker'

* Renamed 'Codeowners Validator' workflow file to easier see what it is for

Before it was named 'sanity.yml', which doesn't make it clear what the workflow is for. Thus 'codeowners_validator.yml' better indicates the purpose of the workflow.

* Made codeowners_validator run on push + restricted it to glotaran orga

Since the workflow depends on 'secrets.OWNERS_VALIDATOR_GITHUB_SECRET' which only exists in the glotaran organization, it is sensible to restrict this workflow only to run in the organization, so forks won't have a failing workflow.