Add failing unit test for parameter parsing from csv
In places where a parameter is used in the context of if , sometimes __bool__ returns a numpy.bool_ which generates a runtime error.
One example is the usage of "if self.dispersion_center:" in the code.

Added unit test simulates this and the next commit will fix it.

At the same time added the test for reading from csv file, which is a requirement to get the numpy.bool_ casting to occur.