Illustrate different ways to refactor tests
See test refactor issue #532

This commit illustrates two possible solution to refactoring the tests:
- Move specification to .yml files so the bots (largely) ignore them / they don't code towards code duplication.
- Make small (yaml) code generator function to build up a model spec.
The other two options are left as an exercise to the reader (to be committed in a follow up commit):
Work with nested dicts (derive from a base model dicts and update/del keys as needed in each tests)
Make a model spec reusable between fitting and simulation. Allowing a model spec to contain e.g. shapes specification which can be ignore during fitting and can be using during simulation.

The latter solution is preferred by the commit's author.
ba2419