Set an COMMIT_EMAIL for create-pr-action in autoupdate-pre-commit-config
The default of that action is '${github.actor}@users.noreply.github.com' which pointed to @jsnel (maybe because he merged the PR back then or because he created the secret) and didn't allow him to review PRs.
The new email points to the github-actions bot, and hopefully fixes this issue.
See: https://github.community/t/github-actions-bot-email-address/17204/4