Replaced 'IPython.display.Code' with custom 'print_yaml_file' function
The problem with 'Code' was, while it was properly syntax highlighted in the Html docs, it doesn't work in jupyterlab (see: https://github.com/ipython/ipython/issues/11747) and the workaround is longer than the markdown wrapping in 'print_yaml_file'. Also, the colormap used by 'pygments' does not adjust or work well with the dark theme.