🩹 Fix missing attribute on old usage for deprecated module
This also fixes the issue that pytest shows a deprecation warning when a test fails since it uses inspect to get all modules '__file__' attribute.
We can't add the '__file__' attribute directly to the module since this would prevent the usage of the module '__getattr__', effectively deactivating the module deprecation.