✨ Added 'set_plugin' functinality to overwrite plugins
This needed plugins to be registered under developer provided names (e.g. ''yaml" and "yml") as well as under their full names (import path) .
For users this will make it easier to overwrite a specific plugin if there is a conflicts or add additional information for reproducebilitys sake.
For plugin developers this makes it easier to implement plugins for existing formats.