♻️ Polish the plugin system (#665)
* ♻️ Reordered dataset and file_name in save_dataset

That way it has the same order (<object_to_save>, <save_path>) as other python functions e.g. in the stdlib

* ♻️ Reordered object and file_name in Project io save functions

That way it has the same order (<object_to_save>, <save_path>) as other python functions e.g. in the stdlib

* 📚 Adjusted plugin how to guide and quickstart to reflect changes

* 🧹 Don't warn if plugin overwrites itselfe (reregistering)

* ✨ Added 'set_plugin' functinality to overwrite plugins

This needed plugins to be registered under developer provided names (e.g. ''yaml" and "yml") as well as under their full names (import path) .
For users this will make it easier to overwrite a specific plugin if there is a conflicts or add additional information for reproducebilitys sake.
For plugin developers this makes it easier to implement plugins for existing formats.

* ✨ Added option to add plugin_names to 'plugin_table' functions

This makes it easier to see which plugin is currently used for which format/model name.

Newly added 'model_plugin_table' with the same functionality.

* 📚 Added documentation about reproducible analysis with plugins

* 🧹 Droped plugin key extending in favour of using set plugin functions

The problem with automatically extending keys is that they are unreliable, e.g. if a user doesn't have conflicting plugins installed keys might not exist.
Also, since the keys are generated when loading the imports the order isn't guaranteed depending on which plugins are installed.

* ✨Changed save and load functions to also accept pathlike input

for file_name/result_path

* 🩹 Set correct stackleven for PluginOverwriteWarning

This way the waring points to the line the registration was called and not to the definition of the base registration function.

* 'Refactored by Sourcery'

* 🩹 Differentiate plugin full names for instanciated plugins

This is needed because instanciated plugins can have different behaviour (e.g. 'yml' and 'yml_str')

* 🧹 Changed save function signature order of builtin plugins

* 🧹 Fixed inconsistent spacing of Parameters an type in docstrings

Co-authored-by: Sourcery AI <>