♻️👌 Removed file fields in ProjectIo like classes and used unified field
If the field isn't an instance of the targetClass it will try to load the class instance from file. This also allows classes like scheme to be initialized with file paths directly.
When dataclasses with file loadable fileds are serialized the objects will be replaced with their source path.

In addition projectIO load and save functions will set the ``source_path`` attribute of the class instances.