🩹 Removed inheriting from 'FileLoadableProtocol'
In python 3.9.7 (currently the only python 3.9 version on conda) there was a bug that ignores the __init__ method of classes that inherit from a Protocol.
Rather than restricting the usable python versions for typing sake, I will file an issue at mypy since the conformity with protocols should be picked up automatically.