🩹🚧 Backport of bugfix
Backport of bugfix #927 discovered in PR #860

It was [found](https://github.com/glotaran/pyglotaran/pull/860#issuecomment-985467588) in #860 that the a_matrix when calculated for result saving was calculated without normalized initial_concentrations while the calculation for the optimization was (correctly) using the normalized version. Fixing this bug in #927 then breaks result consistency checking for some case studies, thus requiring backporting of this bugfix to the 0.4 maintenance branch which we continue to use for result/consistency checking.