🚇 Only run doxygen workflow when requested
Since we most of the time don't make use of the Doxygen generated callgraphs we don't need to always run the workflow.
Thus is created a [github environment](https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment) called `Debug` which needs the approval to run.
That way Doxygen won't run until we approve/request it.