🩹⌨️ Remove usage of 'Number' for typing
With model items and properties we use typing to do type conversion which is incompatible with 'Number' and leads to typing problems.

In [1]: from numbers import Number

In [2]: Number(1)
---------------------------------------------------------------------------
TypeError                                 Traceback (most recent call last)
<ipython-input-2-8247e2ac7336> in <module>
----> 1 Number(1)

TypeError: Number() takes no arguments

Mypy sees float as subclass of int, so passing an int as a flot arg is valid.