🔧Stale bot
Automatically close stale Issues and Pull Requests that tend to accumulate during a project. Ref: https://github.com/marketplace/stale

After a period of inactivity, a label will be applied to mark an issue as stale, and optionally post a comment to notify contributors that the Issue or Pull Request will be closed.

If the Issue or Pull Request is updated, or anyone comments, then the stale label is removed.

If no more activity occurs, the Issue or Pull Request will be automatically closed.