👌 Make PreProcessingPipeline methods return a new instance (#1263)
This changes the behavior of methods on PreProcessingPipeline  to be closer to how pandas or xarray work (object creation instead of in place mutation), which makes it more intuitive and prevents accidental state changes in the pipeline (e.g. executing a cell in a notebook that does not include the instantiation multiple times).