🧰👌 Add TCH ruff rule
Note that we need to ignore this rule for types that are used to define pydantic acctributes or the pydantic models will ony get partially defined.
Errors caused by this look as follows:
pydantic.errors.PydanticUserError: `<pydantic-class>` is not fully defined; you should define `<type-of-attribute>`, then call `<pydantic-class>.model_rebuild()`.