Fix for when the backsweep is a parameter rather than constant
The optimizer would crash

When setting the values in the model scheme from a constant to a parameter, the optimizer would crash with "Cannot determine Numba type of <class 'glotaran.parameter.parameter.Parameter'> " as it tried to use the parameter as a numeric value. Ultimately numba would complain as that's where the value would be first used.

Specifically: `"glotaran\builtin\elements\kinetic\matrix.py", line 27`