🧪🧹 Remove test of upstream schema creation functionality
Rather than testing upstream functionality for an exact match we should test that the schema we generate for a valid spec file validate.
The problem with testing exact schema matching is that it is too tight coupled to the exact syntax of the schema, even so they might be functionally equivalent, leading to failing unit test due to upstream schema reinements.